backend/requirements.txt
Flask
requests
python-dotenv
langchain
langgraph
pydantic
flask-cors
flask-socketio
eventlet
pyjwt
